--- components/tabs.js
@@ -1,491 +1,0 @@
-// @flow
-import React, { useContext } from 'react';
-import { Link, Router } from '../routes';
-import * as Links from './links.js';
-import { poetGenetiveLastName } from './poetname-helpers.js';
-import _ from '../common/translations.js';
-import type { Lang, Poet, Country } from '../common/types.js';
-import CommonData from '../common/commondata.js';
-import LangContext from '../common/LangContext.js';
-import { LoupeSVG, CrossSVG } from './icons.js';
-
-const transitionDuration = '0.2s';
-
-type TabsProps = {
-  items: Array<{ id: string, url: string, title: string, hide?: boolean }>,
-  poet?: Poet,
-  country: Country,
-  lang: Lang,
-  query?: ?string,
-  selected: string,
-};
-type TabsState = {
-  showSearchField: boolean,
-};
-export default class Tabs extends React.Component<TabsProps, TabsState> {
-  searchField: HTMLInputElement;
-  onLoupeClick: (e: Event) => void;
-  onCrossClick: (e: Event) => void;
-  onSubmit: (e: Event) => void;
-  onKeyDown: (e: KeyboardEvent) => void;
-  onFocus: () => void;
-  onBlur: () => void;
-
-  constructor(props: TabsProps) {
-    super(props);
-    const { query } = props;
-    this.state = { showSearchField: query != null && query.length > 0 };
-    this.onLoupeClick = this.onLoupeClick.bind(this);
-    this.onCrossClick = this.onCrossClick.bind(this);
-    this.onSubmit = this.onSubmit.bind(this);
-    this.onKeyDown = this.onKeyDown.bind(this);
-    this.onBlur = this.onBlur.bind(this);
-    this.onFocus = this.onFocus.bind(this);
-  }
-
-  hideSearchField() {
-    this.searchField.value = '';
-    if (window) {
-      window.removeEventListener('keydown', this.onKeyDown, false);
-    }
-    this.setState({ showSearchField: false });
-  }
-
-  componentDidUpdate() {
-    if (this.state.showSearchField) {
-      if (window) {
-        window.addEventListener('keydown', this.onKeyDown, false);
-      }
-      if (document && document.activeElement !== this.searchField) {
-        this.searchField.focus();
-      }
-    }
-  }
-
-  showSeachField() {
-    this.setState({ showSearchField: true });
-  }
-
-  onSubmit(e: Event) {
-    const q = this.searchField.value;
-    const { poet, country, lang } = this.props;
-    let URL = null;
-    if (poet != null && poet.has_texts) {
-      URL = Links.searchURL(lang, q, poet.country, poet.id);
-    } else {
-      URL = Links.searchURL(lang, q, country);
-    }
-    if (document) {
-      this.searchField.blur();
-    }
-    Router.pushRoute(URL);
-    e.preventDefault();
-  }
-
-  onLoupeClick(e: Event) {
-    if (this.state.showSearchField) {
-      const q = this.searchField.value;
-      if (q.length === 0) {
-        // When the text input is empty, the loupe
-        // should simply toggle seach mode.
-        this.hideSearchField();
-      } else {
-        this.onSubmit(e);
-      }
-    } else {
-      this.showSeachField();
-    }
-    e.preventDefault();
-  }
-
-  onCrossClick(e: MouseEvent) {
-    this.hideSearchField();
-    e.preventDefault();
-  }
-
-  onFocus() {
-    window && (window.searchFieldHasFocus = true);
-  }
-
-  onBlur() {
-    window && (window.searchFieldHasFocus = false);
-  }
-
-  onKeyDown(e: KeyboardEvent) {
-    if (e.keyCode === 27) {
-      this.hideSearchField();
-      this.onBlur();
-      e.preventDefault();
-    }
-  }
-
-  render() {
-    const { items, selected, poet, country, lang, query } = this.props;
-    let placeholder = null;
-    if (poet != null && poet.has_texts) {
-      const genetiveLastName = poetGenetiveLastName(poet, lang);
-      placeholder = _('Søg i {genetiveLastName} værker', lang, {
-        genetiveLastName,
-      });
-    } else {
-      if (country === 'dk') {
-        placeholder = _('Søg i Kalliope', lang);
-      } else {
-        const countryData = CommonData.countries.filter(
-          x => x.code === country
-        );
-        if (countryData.length > 0) {
-          const adjective = countryData[0].adjective[lang];
-          placeholder = _('Søg i Kalliopes {adjective} samling', lang, {
-            adjective,
-          });
-        } else {
-          placeholder = _('Søg i Kalliope', lang);
-        }
-      }
-    }
-    const searchField = (
-      <div style={{ display: 'flex' }}>
-        <div style={{ flexGrow: 1 }}>
-          <form onSubmit={this.onSubmit}>
-            <label htmlFor="search-field-id" style={{ display: 'none' }}>
-              Søg
-            </label>
-            <input
-              id="search-field-id"
-              ref={domElement => {
-                if (domElement != null) {
-                  this.searchField = domElement;
-                }
-              }}
-              onFocus={this.onFocus}
-              onBlur={this.onBlur}
-              title={placeholder}
-              defaultValue={query}
-              className="search-field"
-              placeholder={placeholder}
-            />
-          </form>
-        </div>
-        <div
-          className="svg-container"
-          onClick={this.onCrossClick}
-          style={{
-            cursor: 'pointer',
-          }}>
-          <CrossSVG color="black" />
-        </div>
-      </div>
-    );
-
-    const itemsRendered = items
-      .filter(item => !item.hide)
-      .map((item, i) => {
-        const className = item.id === selected ? 'tab selected' : 'tab';
-        return (
-          <div className={className} key={item.url}>
-            <Link route={item.url}>
-              <a>
-                <h2>{item.title}</h2>
-              </a>
-            </Link>
-          </div>
-        );
-      });
-
-    const leftSide = (
-      <div className="leftside">
-        <nav
-          className="tabs"
-          style={{
-            display: this.state.showSearchField ? 'none' : 'block',
-          }}>
-          {itemsRendered}
-        </nav>
-        <div
-          className="searchfield-container"
-          style={{
-            display: this.state.showSearchField ? 'block' : 'none',
-          }}>
-          {searchField}
-        </div>
-      </div>
-    );
-
-    const loupeColor = this.state.showSearchField ? 'black' : '#707070';
-
-    return (
-      <div className="tabs-container">
-        {leftSide}
-        <div className="svg-container" style={{ alignSelf: 'flex-start' }}>
-          <span onClick={this.onLoupeClick} style={{ cursor: 'pointer' }}>
-            <LoupeSVG color={loupeColor} />
-          </span>
-        </div>
-        <style jsx>{`
-          .tabs-container {
-            display: flex;
-            justify-content: space-between;
-            border-bottom: 1px solid black;
-            margin-bottom: 80px;
-            margin-top: 30px;
-            flex-grow: 1;
-          }
-          :global(.svg-container)  {
-            flex-basis: 28px;
-            flex-shrink: 1;
-            align-self: flex-start;
-            transition: flex-basis ${transitionDuration};
-          }
-          :global(.leftside) {
-            width: 100%;
-            padding-right: 10px;
-          }
-          :global(.search-field) {
-            font-size: 32px;
-            line-height: 32px;
-            width: 100%;
-            border: 0;
-            padding: 0;
-            margin: 0;
-            outline: 0;            
-            font-weight: 100;
-            font-family: inherit;
-            transition: font-size ${transitionDuration}, line-height: ${transitionDuration};
-            caret-color: black;
-          }
-          :global(.searchfield-container) {
-            width: 100%;
-            padding-bottom: 16px;
-          }
-          :global(.tabs) > :global(.tab) {
-            display: inline-block;
-            margin-right: 30px;
-            border-bottom: 2px solid transparent;
-            padding-bottom: 15px;
-            transition: margin-right ${transitionDuration};
-          }
-          :global(.tabs) > :global(.tab) :global(h2) {
-            margin: 0;
-            padding: 0;
-            line-height: 32px;
-            font-size: 32px;
-            font-weight: 100;
-            transition: font-size ${transitionDuration}, line-height: ${transitionDuration};
-          }
-          :global(.tabs) > :global(.tab.selected) {
-            border-bottom: 2px solid black;
-          }
-          /*
-          :global(.tabs) > :global(.tab:hover) {
-            border-bottom: 2px solid #888;
-          }
-          */
-          :global(.tabs) :global(.tab.selected a) {
-            color: black;
-          }
-          :global(.tabs) > :global(.tab) :global(a) {
-            color: #707070;
-          }
-          @media (max-width: 850px) {
-            :global(.tabs) > :global(.tab) :global(h2) {
-              font-size: 30px;
-              line-height: 30px;
-            }
-            :global(.svg-container) {
-              flex-basis: 26px;
-            }
-            :global(.tabs) > :global(.tab) {
-              margin-right: 30px;
-            }
-            :global(.search-field) {
-              font-size: 30px;
-              line-height: 30px;
-            }
-          }
-          @media (max-width: 800px) {
-            :global(.tabs) > :global(.tab) :global(h2) {
-              font-size: 28px;
-              line-height: 28px;
-            }
-            :global(.svg-container) {
-              flex-basis: 22px;
-            }
-            :global(.tabs) > :global(.tab) {
-              margin-right: 20px;
-            }
-            :global(.search-field) {
-              font-size: 28px;
-              line-height: 28px;
-            }
-          }
-          @media (max-width: 700px) {
-            :global(.tabs) > :global(.tab) :global(h2) {
-              font-size: 24px;
-              line-height: 24px;
-            }
-            :global(.svg-container) {
-              flex-basis: 20px;
-            }
-            :global(.tabs) > :global(.tab) {
-              margin-right: 15px;
-            }
-            :global(.search-field) {
-              font-size: 24px;
-              line-height: 24px;
-            }
-          }
-          @media (max-width: 600px) {
-            :global(.tabs) > :global(.tab) :global(h2) {
-              font-size: 18px;
-              line-height: 18px;
-            }
-            :global(.svg-container) {
-              flex-basis: 16px;
-            }
-            :global(.tabs) > :global(.tab) {
-              margin-right: 15px;
-            }
-            :global(.search-field) {
-              font-size: 18px;
-              line-height: 18px;
-            }
-          }
-
-          @media (max-width: 480px) {
-            :global(.tabs) > :global(.tab) {
-              margin-right: 10px;
-            }
-            :global(.tabs) > :global(.tab) :global(h2) {
-              font-size: 12px !important;
-              line-height: 12px;
-            }
-            :global(.svg-container) {
-              flex-basis: 14px;
-            }
-            :global(.search-field) {
-              font-size: 12px;
-              line-height: 12px;
-            }
-          }
-
-          @media (max-width: 320px) {
-            :global(.tabs) > :global(.tab) {
-              margin-right: 4px;
-            }
-            :global(.tabs) > :global(.tab) :global(h2) {
-              font-size: 10px !important;
-              line-height: 10px;
-            }
-            :global(.svg-container) {
-              flex-basis: 10px;
-            }
-            :global(.search-field) {
-              font-size: 10px;
-              line-height: 10px;
-            }
-          }
-
-          @media print {
-            .tabs-container {
-              display: none;
-            }
-          }
-        `}</style>
-      </div>
-    );
-  }
-}
-
-export const poetTabs = (poet: Poet) => {
-  const lang = useContext(LangContext);
-
-  return [
-    {
-      id: 'works',
-      title: _('Værker', lang),
-      hide: !poet.has_works && !poet.has_artwork,
-      url: Links.worksURL(lang, poet.id),
-    },
-    {
-      id: 'titles',
-      title: _('Digttitler', lang),
-      hide: !poet.has_poems,
-      url: Links.textsURL(lang, poet.id, 'titles'),
-    },
-    {
-      id: 'first',
-      title: _('Førstelinjer', lang),
-      hide: !poet.has_poems,
-      url: Links.textsURL(lang, poet.id, 'first'),
-    },
-    {
-      id: 'mentions',
-      title: _('Henvisninger', lang),
-      hide: !poet.has_mentions,
-      url: Links.mentionsURL(lang, poet.id),
-    },
-    {
-      id: 'bio',
-      title: _('Biografi', lang),
-      hide: !poet.has_biography,
-      url: Links.bioURL(lang, poet.id),
-    },
-  ];
-};
-
-export const kalliopeTabs = () => {
-  const lang = useContext(LangContext);
-  return [
-    { id: 'index', title: 'Kalliope', url: Links.frontPageURL(lang) },
-    {
-      id: 'poets',
-      title: _('Digtere', lang),
-      url: Links.poetsURL(lang, 'name'),
-    },
-    /*
-      {
-        id: 'poems',
-        title: _('Digte', lang),
-        url: Links.allTextsURL(lang, 'dk', 'titles', 'A'),
-      },
-      */
-    {
-      id: 'keywords',
-      title: _('Nøgleord', lang),
-      url: Links.keywordsURL(lang),
-    },
-    //{ id: 'dictionary', title: 'Ordbog', url: Links.dictionaryURL(lang) },
-    {
-      id: 'about',
-      title: _('Om', lang),
-      url: Links.aboutURL(lang, 'kalliope'),
-    },
-  ];
-};
-
-type KalliopeTabsProps = {
-  country?: Country,
-  query?: ?string,
-  selected:
-    | 'index'
-    | 'poets'
-    | 'keywords'
-    | 'dictionary'
-    | 'about'
-    | 'search'
-    | 'museum',
-};
-export const KalliopeTabs = (props: KalliopeTabsProps) => {
-  const { selected, country, query } = props;
-  const lang = useContext(LangContext);
-  return (
-    <Tabs
-      items={kalliopeTabs()}
-      selected={selected}
-      lang={lang}
-      country={country || 'dk'}
-      query={query}
-    />
-  );
-};
