+++ pages/keyword.js
@@ -7,8 +7,7 @@ import Page from '../components/page.js';
 import { kalliopeCrumbs } from '../components/breadcrumbs.js';
 import SidebarSplit from '../components/sidebarsplit.js';
 import LangSelect from '../components/langselect';
-import { kalliopeTabs } from '../components/tabs.js';
-import Heading from '../components/heading.js';
+import { kalliopeTabs } from '../components/menu.js';
 import SubHeading from '../components/subheading.js';
 import PoetName from '../components/poetname.js';
 import TextName from '../components/textname.js';
