+++ pages/keywords.js
@@ -6,7 +6,7 @@ import Page from '../components/page.js';
 import * as Links from '../components/links';
 import { kalliopeCrumbs } from '../components/breadcrumbs.js';
 import LangSelect from '../components/langselect.js';
-import { kalliopeTabs } from '../components/tabs.js';
+import { kalliopeTabs } from '../components/menu.js';
 import PoetName from '../components/poetname.js';
 import SectionedList from '../components/sectionedlist.js';
 import * as Sorting from '../common/sorting.js';
