+++ pages/mentions.js
@@ -6,8 +6,7 @@ import { poetCrumbsWithTitle } from '../components/breadcrumbs.js';
 import _ from '../common/translations.js';
 import { Link } from '../routes';
 import LangSelect from '../components/langselect';
-import { poetTabs } from '../components/tabs.js';
-import Heading from '../components/heading.js';
+import { poetTabs } from '../components/menu.js';
 import PoetName from '../components/poetname.js';
 import { poetNameString } from '../components/poetname-helpers.js';
 import TextContent from '../components/textcontent.js';
