+++ pages/museum.js
@@ -7,8 +7,7 @@ import Page from '../components/page.js';
 import Main from '../components/main.js';
 import { kalliopeCrumbs } from '../components/breadcrumbs.js';
 import LangSelect from '../components/langselect';
-import { kalliopeTabs } from '../components/tabs.js';
-import Heading from '../components/heading.js';
+import { kalliopeTabs } from '../components/menu.js';
 import PoetName from '../components/poetname.js';
 import WorkName from '../components/workname.js';
 import PicturesGrid from '../components/picturesgrid.js';
