+++ pages/poets-looks.js
@@ -6,8 +6,7 @@ import Page from '../components/page.js';
 import { Link } from '../routes';
 import * as Links from '../components/links';
 import LangSelect from '../components/langselect.js';
-import Tabs from '../components/tabs.js';
-import Heading from '../components/heading.js';
+import Tabs from '../components/menu.js';
 import PoetName from '../components/poetname.js';
 import { poetNameString } from '../components/poetname-helpers.js';
 import Picture from '../components/picture.js';
