+++ pages/poets.js
@@ -10,8 +10,7 @@ import { kalliopeCrumbs } from '../components/breadcrumbs.js';
 import LangSelect from '../components/langselect.js';
 import LangContext from '../common/LangContext.js';
 import CountryPicker from '../components/countrypicker.js';
-import Tabs from '../components/tabs.js';
-import Heading from '../components/heading.js';
+import Tabs from '../components/menu.js';
 import PoetName from '../components/poetname.js';
 import SectionedList from '../components/sectionedlist.js';
 import * as Sorting from '../common/sorting.js';
