+++ pages/text.js
@@ -8,8 +8,7 @@ import Main from '../components/main.js';
 import { textCrumbs } from '../components/breadcrumbs.js';
 import SidebarSplit from '../components/sidebarsplit.js';
 import LangSelect from '../components/langselect';
-import { poetTabs } from '../components/tabs.js';
-import Heading from '../components/heading.js';
+import { poetTabs } from '../components/menu.js';
 import SubHeading from '../components/subheading.js';
 import Stack from '../components/stack.js';
 import WrapNonEmpty from '../components/wrapnonempty.js';
