+++ pages/texts.js
@@ -4,11 +4,10 @@ import React from 'react';
 import Page from '../components/page.js';
 import { poetCrumbsWithTitle } from '../components/breadcrumbs.js';
 import LangSelect from '../components/langselect.js';
-import Heading from '../components/heading.js';
 import PoetName from '../components/poetname.js';
 import { poetNameString } from '../components/poetname-helpers.js';
 import WorkName from '../components/workname.js';
-import { poetTabs } from '../components/tabs.js';
+import { poetTabs } from '../components/menu.js';
 import SectionedList from '../components/sectionedlist.js';
 import * as Links from '../components/links.js';
 import * as Sorting from '../common/sorting.js';
