+++ pages/work.js
@@ -7,8 +7,7 @@ import Main from '../components/main.js';
 import { workCrumbs } from '../components/breadcrumbs.js';
 import SidebarSplit from '../components/sidebarsplit.js';
 import LangSelect from '../components/langselect.js';
-import { poetTabs } from '../components/tabs.js';
-import Heading from '../components/heading.js';
+import { poetTabs } from '../components/menu.js';
 import SubHeading from '../components/subheading.js';
 import PoetName from '../components/poetname.js';
 import { poetNameString } from '../components/poetname-helpers.js';
