+++ pages/works.js
@@ -6,8 +6,7 @@ import { Link, Router } from '../routes';
 import Page from '../components/page.js';
 import { worksCrumbs } from '../components/breadcrumbs.js';
 import LangSelect from '../components/langselect';
-import { poetTabs } from '../components/tabs.js';
-import Heading from '../components/heading.js';
+import { poetTabs } from '../components/menu.js';
 import PoetName from '../components/poetname.js';
 import { poetNameString } from '../components/poetname-helpers.js';
 import PicturesGrid from '../components/picturesgrid.js';
